+++ Backend/fcsp-webapi/FCSP.DTOs/SetServiceAmount/SetServiceAmountDTOs.cs
@@ -11,7 +11,7 @@ namespace FCSP.DTOs.SetServiceAmount
         public long ServiceId { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime? EndDate { get; set; }
-        public float Amount { get; set; }
+        public int Amount { get; set; }
         public int Status { get; set; }
         public string ServiceName { get; set; } = null!;
         public DateTime CreatedAt { get; set; }
@@ -23,7 +23,7 @@ namespace FCSP.DTOs.SetServiceAmount
         public long ServiceId { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime? EndDate { get; set; }
-        public float Amount { get; set; }
+        public int Amount { get; set; }
         public int Status { get; set; }
     }
 
@@ -31,7 +31,7 @@ namespace FCSP.DTOs.SetServiceAmount
     {
         public long Id { get; set; }
         public long ServiceId { get; set; }
-        public float Amount { get; set; }
+        public int Amount { get; set; }
         public DateTime CreatedAt { get; set; }
     }
 
@@ -39,14 +39,14 @@ namespace FCSP.DTOs.SetServiceAmount
     {
         public long Id { get; set; }
         public DateTime? EndDate { get; set; }
-        public float Amount { get; set; }
+        public int Amount { get; set; }
         public int Status { get; set; }
     }
 
     public class UpdateSetServiceAmountResponse
     {
         public long Id { get; set; }
-        public float Amount { get; set; }
+        public int Amount { get; set; }
         public int Status { get; set; }
         public DateTime UpdatedAt { get; set; }
     }
