+++ Backend/fcsp-webapi/FCSP.Services/ServiceService/IServiceService.cs
@@ -10,7 +10,7 @@ namespace FCSP.Services.ServiceService
         Task<BaseResponseModel<AddServiceResponse>> AddService(AddServiceRequest request);
         Task<BaseResponseModel<UpdateServiceResponse>> UpdateService(UpdateServiceRequest request);
         Task<BaseResponseModel<DeleteServiceResponse>> DeleteService(DeleteServiceRequest request);
-        Task<float?> GetServicePriceAsync(long serviceId);
+        Task<int?> GetServicePriceAsync(long serviceId);
         Task<BaseResponseModel<List<ServiceResponseDto>>> GetServicesByManufacturerId(long manufacturerId); // New method
     }
 }
